% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demcon-package.R
\docType{package}
\name{demcon-package}
\alias{demcon}
\alias{demcon-package}
\title{demcon: Interfacing with Popular Polity, Institutional, and Constitutional Datasets}
\description{
An open-source toolkit developed by ISciences, LLC and the DANTE Project that is intended for learning, accessing, pre-processing, and visualizing popular political, institutional, and constitutional datasets. ‘demcon' aims to reduce barriers to entry in political science research by automating common acquisition and pre-processing procedures. This package particularly focuses on the V-Dem dataset (\url{https://www.v-dem.net/vdemds.html}), and provides adaptations of methods presented in Fjelde, H., Knutsen, C. H. & Nygård, H. M. 2021, \doi{10.1093/isq/sqaa076}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dante-sttr.gitlab.io/demcon/index.html}
  \item \url{https://gitlab.com/dante-sttr/demcon}
  \item \url{https://www.isciences.com}
  \item Report bugs at \url{https://gitlab.com/dante-sttr/demcon/-/issues}
}

}
\author{
\strong{Maintainer}: Joshua Brinks \email{jbrinks@isciences.com}

}
\keyword{internal}
