% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-functions.R
\name{range01}
\alias{range01}
\title{Standardize a Variable to 0-1}
\usage{
range01(x)
}
\arguments{
\item{x}{A vector of numeric values.}
}
\value{
A standardized numeric vector scaled to 0-1.
}
\description{
Standardize a vector to 0-1 using the cumulative distribution function
of the normal distribution.
}
\examples{

nums<-rnorm(50, 6.5, 3)

nums<-demcon::range01(nums)

}
